backend/requirements.txt
Django>=4.2
 djangorestframework
 djangorestframework-simplejwt
 channels
 uvicorn
 asgiref
 python-dotenv
 sentence-transformers
 faiss-cpu
 transformers
 torch
 pytest
 pytest-django
 django-cors-headers
